only 2 tests failing

+++ src/main/java/app/models/ModelObject.java
@@ -1,6 +1,6 @@
 package app.models;
 
 public interface ModelObject {
-    Long getId();
-    void setId(Long id);
+    int getId();
+    void setId(int id);
 }
